Split out formatting and comparison methods
More smaller traits helps address the monolith of code in DateTimeTrait.
I've chosen names that attempt to pair up the test suites with code
bearing traits.
