+++ src/DateTimeTrait.php
@@ -3,7 +3,8 @@
  * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
  *
  * Licensed under The MIT License
- * Redistributions of files must retain the above copyright notice.
+ * Redistributions of files must retain the above copyright notice. Provides various operator methods for datetime 
+ * objects.
  *
  * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
  * @copyright     Copyright (c) Brian Nesbitt <brian@nesbot.com>
@@ -12,6 +13,8 @@
  */
 namespace Cake\Chronos;
 
+use Cake\Chronos\ComparisonTrait;
+use Cake\Chronos\FormattingTrait;
 use DatePeriod;
 use DateTime;
 use DateTimeInterface;
@@ -23,6 +26,8 @@ use InvalidArgumentException;
  */
 trait DateTimeTrait
 {
+    use ComparisonTrait;
+    use FormattingTrait;
 
     /**
      * Names of days of the week.
@@ -662,312 +667,6 @@ trait DateTimeTrait
         }
 
         return false;
-    }
-
-    /**
-     * Reset the format used to the default when type juggling a ChronosInterface instance to a string
-     *
-     * @return void
-     */
-    public static function resetToStringFormat()
-    {
-        static::setToStringFormat(ChronosInterface::DEFAULT_TO_STRING_FORMAT);
-    }
-
-    /**
-     * Set the default format used when type juggling a ChronosInterface instance to a string
-     *
-     * @param string $format The format to use in future __toString() calls.
-     * @return void
-     */
-    public static function setToStringFormat($format)
-    {
-        static::$toStringFormat = $format;
-    }
-
-    /**
-     * Format the instance as a string using the set format
-     *
-     * @return string
-     */
-    public function __toString()
-    {
-        return $this->format(static::$toStringFormat);
-    }
-
-    /**
-     * Format the instance as date
-     *
-     * @return string
-     */
-    public function toDateString()
-    {
-        return $this->format('Y-m-d');
-    }
-
-    /**
-     * Format the instance as a readable date
-     *
-     * @return string
-     */
-    public function toFormattedDateString()
-    {
-        return $this->format('M j, Y');
-    }
-
-    /**
-     * Format the instance as time
-     *
-     * @return string
-     */
-    public function toTimeString()
-    {
-        return $this->format('H:i:s');
-    }
-
-    /**
-     * Format the instance as date and time
-     *
-     * @return string
-     */
-    public function toDateTimeString()
-    {
-        return $this->format('Y-m-d H:i:s');
-    }
-
-    /**
-     * Format the instance with day, date and time
-     *
-     * @return string
-     */
-    public function toDayDateTimeString()
-    {
-        return $this->format('D, M j, Y g:i A');
-    }
-
-    /**
-     * Format the instance as ATOM
-     *
-     * @return string
-     */
-    public function toAtomString()
-    {
-        return $this->format(DateTime::ATOM);
-    }
-
-    /**
-     * Format the instance as COOKIE
-     *
-     * @return string
-     */
-    public function toCookieString()
-    {
-        return $this->format(DateTime::COOKIE);
-    }
-
-    /**
-     * Format the instance as ISO8601
-     *
-     * @return string
-     */
-    public function toIso8601String()
-    {
-        return $this->format(DateTime::ISO8601);
-    }
-
-    /**
-     * Format the instance as RFC822
-     *
-     * @return string
-     */
-    public function toRfc822String()
-    {
-        return $this->format(DateTime::RFC822);
-    }
-
-    /**
-     * Format the instance as RFC850
-     *
-     * @return string
-     */
-    public function toRfc850String()
-    {
-        return $this->format(DateTime::RFC850);
-    }
-
-    /**
-     * Format the instance as RFC1036
-     *
-     * @return string
-     */
-    public function toRfc1036String()
-    {
-        return $this->format(DateTime::RFC1036);
-    }
-
-    /**
-     * Format the instance as RFC1123
-     *
-     * @return string
-     */
-    public function toRfc1123String()
-    {
-        return $this->format(DateTime::RFC1123);
-    }
-
-    /**
-     * Format the instance as RFC2822
-     *
-     * @return string
-     */
-    public function toRfc2822String()
-    {
-        return $this->format(DateTime::RFC2822);
-    }
-
-    /**
-     * Format the instance as RFC3339
-     *
-     * @return string
-     */
-    public function toRfc3339String()
-    {
-        return $this->format(DateTime::RFC3339);
-    }
-
-    /**
-     * Format the instance as RSS
-     *
-     * @return string
-     */
-    public function toRssString()
-    {
-        return $this->format(DateTime::RSS);
-    }
-
-    /**
-     * Format the instance as W3C
-     *
-     * @return string
-     */
-    public function toW3cString()
-    {
-        return $this->format(DateTime::W3C);
-    }
-
-    /**
-     * Determines if the instance is equal to another
-     *
-     * @param ChronosInterface $dt The instance to compare with.
-     * @return bool
-     */
-    public function eq(ChronosInterface $dt)
-    {
-        return $this == $dt;
-    }
-
-    /**
-     * Determines if the instance is not equal to another
-     *
-     * @param ChronosInterface $dt The instance to compare with.
-     * @return bool
-     */
-    public function ne(ChronosInterface $dt)
-    {
-        return !$this->eq($dt);
-    }
-
-    /**
-     * Determines if the instance is greater (after) than another
-     *
-     * @param ChronosInterface $dt The instance to compare with.
-     * @return bool
-     */
-    public function gt(ChronosInterface $dt)
-    {
-        return $this > $dt;
-    }
-
-    /**
-     * Determines if the instance is greater (after) than or equal to another
-     *
-     * @param ChronosInterface $dt The instance to compare with.
-     * @return bool
-     */
-    public function gte(ChronosInterface $dt)
-    {
-        return $this >= $dt;
-    }
-
-    /**
-     * Determines if the instance is less (before) than another
-     *
-     * @param ChronosInterface $dt The instance to compare with.
-     * @return bool
-     */
-    public function lt(ChronosInterface $dt)
-    {
-        return $this < $dt;
-    }
-
-    /**
-     * Determines if the instance is less (before) or equal to another
-     *
-     * @param ChronosInterface $dt The instance to compare with.
-     * @return bool
-     */
-    public function lte(ChronosInterface $dt)
-    {
-        return $this <= $dt;
-    }
-
-    /**
-     * Determines if the instance is between two others
-     *
-     * @param ChronosInterface $dt1 The instance to compare with.
-     * @param ChronosInterface $dt2 The instance to compare with.
-     * @param bool $equal Indicates if a > and < comparison should be used or <= or >=
-     * @return bool
-     */
-    public function between(ChronosInterface $dt1, ChronosInterface $dt2, $equal = true)
-    {
-        if ($dt1->gt($dt2)) {
-            $temp = $dt1;
-            $dt1 = $dt2;
-            $dt2 = $temp;
-        }
-
-        if ($equal) {
-            return $this->gte($dt1) && $this->lte($dt2);
-        } else {
-            return $this->gt($dt1) && $this->lt($dt2);
-        }
-    }
-
-    /**
-     * Get the minimum instance between a given instance (default now) and the current instance.
-     *
-     * @param ChronosInterface $dt The instance to compare with.
-     * @return static
-     */
-    public function min(ChronosInterface $dt = null)
-    {
-        $dt = ($dt === null) ? static::now($this->tz) : $dt;
-
-        return $this->lt($dt) ? $this : $dt;
-    }
-
-    /**
-     * Get the maximum instance between a given instance (default now) and the current instance.
-     *
-     * @param ChronosInterface $dt The instance to compare with.
-     * @return static
-     */
-    public function max(ChronosInterface $dt = null)
-    {
-        $dt = ($dt === null) ? static::now($this->tz) : $dt;
-
-        return $this->gt($dt) ? $this : $dt;
     }
 
     /**
